OpenCL Single and Multi Coverage and Optimization (#70)
* Moving agnostic methods back into QInterface

* Moving agnostic public methods into interface

* M() operator OCL coverage

* ApplyM bug fix; INCS/DECS OCL coverage

* INCBCD/DECBCD OCL coverage

* INCBCDC/DECBCDC OCL coverage

* INCSC/DECSC coverage

* PhaseFlip OCL coverage

* ZeroPhaseFlip OCL coverage

* CPhaseFlipIfLess OCL coverage

* Removing dependence of QEngineOCL on QEngineCPU (full OpenCL coverage)

* Debugging OpenCL with map/unmap removed

* Debugging OpenCL

* Debugging OpenCL

* Passes all unit tests

* Optimizing SetPermutation

* Optimizing ProbAll

* Optimizing ProbAll and NormalizeState

* Read/write flags

* Should unmap before deallocating buffer

* Fixing bad array init

* Revert to intended default device

* Read/write flage

* Optimizing multi-processor engine

* Optimizing multi-processor engine

* Optimizing multi-processor engine

* Fixing workgroup size

* Fixing workgroup size

* Fixing workgroup size

* Fixing group size/count

* Seg faults

* Seg fault in benchmarks

* Reverting multi engine

* Reverting multi engine

* Complete kernels before reusing buffers

* Fixed benchmark segfault

* Per review

* Formatting/cleaning before merge

* Removing BCD unit tests

* Finalizing PR
