+++ include/common/oclengine.hpp
@@ -52,11 +52,25 @@ enum OCLAPI {
     OCL_API_DEC,
     OCL_API_INCC,
     OCL_API_DECC,
+    OCL_API_INCS,
+    OCL_API_DECS,
+    OCL_API_INCSC_1,
+    OCL_API_DECSC_1,
+    OCL_API_INCSC_2,
+    OCL_API_DECSC_2,
+    OCL_API_INCBCD,
+    OCL_API_DECBCD,
+    OCL_API_INCBCDC,
+    OCL_API_DECBCDC,
     OCL_API_INDEXEDLDA,
     OCL_API_INDEXEDADC,
     OCL_API_INDEXEDSBC,
     OCL_API_NORMALIZE,
     OCL_API_UPDATENORM,
+    OCL_API_APPLYM,
+    OCL_API_PHASEFLIP,
+    OCL_API_ZEROPHASEFLIP,
+    OCL_API_CPHASEFLIPIFLESS
 };
 
 class OCLDeviceCall {
