+++ include/qengine_opencl.hpp
@@ -16,7 +16,7 @@
 #error OpenCL has not been enabled
 #endif
 
-#include "qengine_cpu.hpp"
+#include "qinterface.hpp"
 
 namespace Qrack {
 
@@ -29,8 +29,9 @@ class QEngineOCL;
 typedef std::shared_ptr<QEngineOCL> QEngineOCLPtr;
 
 /** OpenCL enhanced QEngineCPU implementation. */
-class QEngineOCL : public QEngineCPU {
+class QEngineOCL : public QInterface {
 protected:
+    complex* stateVec;
     int deviceID;
     DeviceContextPtr device_context;
     cl::CommandQueue queue;
@@ -41,7 +42,11 @@ protected:
     cl::Buffer cmplxBuffer;
     cl::Buffer ulongBuffer;
     cl::Buffer nrmBuffer;
-    cl::Buffer maxBuffer;
+    real1* nrmArray;
+    size_t nrmGroupCount;
+    size_t nrmGroupSize;
+
+    virtual void ApplyM(bitCapInt qPower, bool result, complex nrm);
 
 public:
     /**
@@ -55,17 +60,14 @@ public:
      * Qrack::QEngineOCLMulti object, in which case this Qrack::QEngineOCL object might not contain the amplitude of the
      * overall permutation state of the combined object.
      */
-    QEngineOCL(bitLenInt qBitCount, bitCapInt initState, std::shared_ptr<std::default_random_engine> rgp = nullptr,
-        int devID = -1, bool partialInit = false)
-        : QEngineCPU(qBitCount, initState, rgp, complex(-999.0, -999.0), partialInit)
-    {
-        InitOCL(devID);
-    }
 
-    QEngineOCL(QEngineOCLPtr toCopy)
-        : QEngineCPU(toCopy)
+    QEngineOCL(bitLenInt qBitCount, bitCapInt initState, std::shared_ptr<std::default_random_engine> rgp = nullptr,
+        int devID = -1, bool partialInit = false, complex phaseFac = complex(-999.0, -999.0));
+    QEngineOCL(QEngineOCLPtr toCopy);
+    ~QEngineOCL()
     {
-        InitOCL(toCopy->deviceID);
+        delete[] stateVec;
+        delete[] nrmArray;
     }
 
     virtual void SetQubitCount(bitLenInt qb)
@@ -74,25 +76,55 @@ public:
         maxQPower = 1 << qubitCount;
     }
 
+    virtual void EnableNormalize(bool doN) { doNormalize = doN; }
+    virtual real1 GetNorm(bool update = true)
+    {
+        if (update) {
+            UpdateRunningNorm();
+        }
+        return runningNorm;
+    }
+    virtual void SetNorm(real1 n) { runningNorm = n; }
+
+    // CL_MAP_READ = (1 << 0); CL_MAP_WRITE = (1 << 1);
+    virtual void LockSync(cl_int flags = (CL_MAP_READ | CL_MAP_WRITE));
+    virtual void UnlockSync();
     virtual complex* GetStateVector() { return stateVec; }
+    virtual void SetPermutation(bitCapInt perm);
+    virtual void CopyState(QInterfacePtr orig);
+    virtual real1 ProbAll(bitCapInt fullRegister);
 
     /* Operations that have an improved implementation. */
-    virtual void Swap(bitLenInt qubit1, bitLenInt qubit2); // Inherited overload
+    using QInterface::X;
+    virtual void X(bitLenInt start, bitLenInt length);
+    using QInterface::Swap;
     virtual void Swap(bitLenInt start1, bitLenInt start2, bitLenInt length);
-    using QEngineCPU::Cohere;
+
     virtual bitLenInt Cohere(QEngineOCLPtr toCopy);
     virtual bitLenInt Cohere(QInterfacePtr toCopy) { return Cohere(std::dynamic_pointer_cast<QEngineOCL>(toCopy)); }
-    using QEngineCPU::Decohere;
     virtual void Decohere(bitLenInt start, bitLenInt length, QInterfacePtr dest);
     virtual void Dispose(bitLenInt start, bitLenInt length);
-    using QEngineCPU::X;
-    virtual void X(bitLenInt start, bitLenInt length);
+
     virtual void ROL(bitLenInt shift, bitLenInt start, bitLenInt length);
     virtual void ROR(bitLenInt shift, bitLenInt start, bitLenInt length);
+
     virtual void INC(bitCapInt toAdd, bitLenInt start, bitLenInt length);
     virtual void DEC(bitCapInt toSub, bitLenInt start, bitLenInt length);
     virtual void INCC(bitCapInt toAdd, bitLenInt start, bitLenInt length, bitLenInt carryIndex);
     virtual void DECC(bitCapInt toSub, bitLenInt start, bitLenInt length, bitLenInt carryIndex);
+    virtual void INCS(bitCapInt toAdd, bitLenInt start, bitLenInt length, bitLenInt carryIndex);
+    virtual void DECS(bitCapInt toSub, bitLenInt start, bitLenInt length, bitLenInt carryIndex);
+    virtual void INCSC(
+        bitCapInt toAdd, bitLenInt start, bitLenInt length, bitLenInt overflowIndex, bitLenInt carryIndex);
+    virtual void DECSC(
+        bitCapInt toAdd, bitLenInt start, bitLenInt length, bitLenInt overflowIndex, bitLenInt carryIndex);
+    virtual void INCSC(bitCapInt toAdd, bitLenInt start, bitLenInt length, bitLenInt carryIndex);
+    virtual void DECSC(bitCapInt toAdd, bitLenInt start, bitLenInt length, bitLenInt carryIndex);
+    virtual void INCBCD(bitCapInt toAdd, bitLenInt start, bitLenInt length);
+    virtual void DECBCD(bitCapInt toAdd, bitLenInt start, bitLenInt length);
+    virtual void INCBCDC(bitCapInt toAdd, bitLenInt start, bitLenInt length, bitLenInt carryIndex);
+    virtual void DECBCDC(bitCapInt toSub, bitLenInt start, bitLenInt length, bitLenInt carryIndex);
+
     virtual bitCapInt IndexedLDA(bitLenInt indexStart, bitLenInt indexLength, bitLenInt valueStart,
         bitLenInt valueLength, unsigned char* values);
     virtual bitCapInt IndexedADC(bitLenInt indexStart, bitLenInt indexLength, bitLenInt valueStart,
@@ -102,8 +134,14 @@ public:
 
     virtual real1 Prob(bitLenInt qubit);
 
+    virtual void PhaseFlip();
+    virtual void ZeroPhaseFlip(bitLenInt start, bitLenInt length);
+    virtual void CPhaseFlipIfLess(bitCapInt greaterPerm, bitLenInt start, bitLenInt length, bitLenInt flagIndex);
+
     virtual int GetDeviceID() { return deviceID; }
     virtual void SetDevice(const int& dID);
+
+    virtual void SetQuantumState(complex* inputState);
 
     virtual void NormalizeState(real1 nrm = -999.0);
     virtual void UpdateRunningNorm();
@@ -112,21 +150,29 @@ protected:
     static const int BCI_ARG_LEN = 10;
 
     void InitOCL(int devID);
-    void ReInitOCL();
-    void ResetStateVec(complex* nStateVec);
     void ResetStateVec(complex* nStateVec, BufferPtr nStateBuffer);
+    virtual complex* AllocStateVec(bitCapInt elemCount);
 
     void DecohereDispose(bitLenInt start, bitLenInt length, QEngineOCLPtr dest);
-    void DispatchCall(OCLAPI api_call, bitCapInt (&bciArgs)[BCI_ARG_LEN], complex* nVec = NULL,
-        unsigned char* values = NULL, bitCapInt valuesLength = 0);
+    void DispatchCall(
+        OCLAPI api_call, bitCapInt (&bciArgs)[BCI_ARG_LEN], unsigned char* values = NULL, bitCapInt valuesLength = 0);
 
     void Apply2x2(bitCapInt offset1, bitCapInt offset2, const complex* mtrx, const bitLenInt bitCount,
         const bitCapInt* qPowersSorted, bool doCalcNorm);
 
     /* Utility functions used by the operations above. */
     void ROx(OCLAPI api_call, bitLenInt shift, bitLenInt start, bitLenInt length);
-    void INT(OCLAPI api_call, bitCapInt toAdd, const bitLenInt inOutStart, const bitLenInt length);
-    void INTC(OCLAPI api_call, bitCapInt toAdd, const bitLenInt inOutStart, const bitLenInt length,
+    void INT(OCLAPI api_call, bitCapInt toMod, const bitLenInt inOutStart, const bitLenInt length);
+    void INTC(OCLAPI api_call, bitCapInt toMod, const bitLenInt inOutStart, const bitLenInt length,
+        const bitLenInt carryIndex);
+    void INTS(OCLAPI api_call, bitCapInt toMod, const bitLenInt inOutStart, const bitLenInt length,
+        const bitLenInt overflowIndex);
+    void INTSC(OCLAPI api_call, bitCapInt toMod, const bitLenInt inOutStart, const bitLenInt length,
+        const bitLenInt carryIndex);
+    void INTSC(OCLAPI api_call, bitCapInt toMod, const bitLenInt inOutStart, const bitLenInt length,
+        const bitLenInt overflowIndex, const bitLenInt carryIndex);
+    void INTBCD(OCLAPI api_call, bitCapInt toMod, const bitLenInt inOutStart, const bitLenInt length);
+    void INTBCDC(OCLAPI api_call, bitCapInt toMod, const bitLenInt inOutStart, const bitLenInt length,
         const bitLenInt carryIndex);
 
     bitCapInt OpIndexed(OCLAPI api_call, bitCapInt carryIn, bitLenInt indexStart, bitLenInt indexLength,
