+++ src/qengine/state.cpp
@@ -60,13 +60,23 @@ QEngineCPU::QEngineCPU(bitLenInt qBitCount, bitCapInt initState, std::shared_ptr
 }
 
 QEngineCPU::QEngineCPU(QEngineCPUPtr toCopy)
-    : QInterface(toCopy->qubitCount, toCopy->rand_generator)
-    , doNormalize(toCopy->doNormalize)
+    : QInterface(toCopy->qubitCount, toCopy->rand_generator, toCopy->doNormalize)
+    , stateVec(NULL)
 {
+    SetConcurrencyLevel(std::thread::hardware_concurrency());
+    stateVec = AllocStateVec(maxQPower);
     CopyState(toCopy);
 }
 
-complex* QEngineCPU::GetState() { return stateVec; }
+complex* QEngineCPU::GetStateVector() { return stateVec; }
+
+void QEngineCPU::SetPermutation(bitCapInt perm)
+{
+    std::fill(stateVec, stateVec + maxQPower, complex(0.0, 0.0));
+    real1 angle = Rand() * 2.0 * M_PI;
+    stateVec[perm] = complex(cos(angle), sin(angle));
+    runningNorm = 1.0;
+}
 
 void QEngineCPU::CopyState(QInterfacePtr orig)
 {
@@ -75,7 +85,7 @@ void QEngineCPU::CopyState(QInterfacePtr orig)
     ResetStateVec(AllocStateVec(maxQPower));
 
     QEngineCPUPtr src = std::dynamic_pointer_cast<QEngineCPU>(orig);
-    std::copy(src->GetState(), src->GetState() + (1 << (src->GetQubitCount())), stateVec);
+    std::copy(src->stateVec, src->stateVec + (1 << (src->qubitCount)), stateVec);
 }
 
 void QEngineCPU::ResetStateVec(complex* nStateVec)
